cmake: fully apply target syntax to swig
Previously a global variable was used to define
the runtime swig include directory. Also include_directories
and link_libraries was used to define the includes.

Now the runtime_swig dependency is used to define the swig include
directory and the transient link flags from the C++ library is used
to link to correct libraries.